MNT: Moved Plotter into the controlsui package
